libcrush: get all pps values for a Ceph pool
For the Ceph pool 0 that has 1024 PGs, each PG will be mapped to the
number of OSDs required by the replication count. The number given to
crush_do_rule for that mapping is named the pps in the code but never
exposed to the user. It roughly is pool hash(number + pg number), i.e
hash(0.3), hash(0.4) etc.

The pool_pps returns the entire list of values for a given pool and a
given number of PGs. These values can then be given to the crush_do_rule
function with the cluster crushmap to get the exact layout of the PG,
from the Ceph point of view.

Signed-off-by: Loic Dachary <loic@dachary.org>