ceph: always write all tunables
CrushWrapper does not write the tunables that have some ancient values.
This can create a problem when using the crushmap on old clusters that
have different defaults. For instance chooseleaf_stable, if omitted will
default to 1 in Luminous and to 0 in Jewel.

The list of tunables depends on the Ceph version. Before hammer
chooseleaf_stable did not exist.

Signed-off-by: Loic Dachary <loic@dachary.org>